App/requirements.txt
requests
RPi.GPIO
schedule
Flask